Extract code for ReplaceFileButton to own component
